Restructure and working tests
